requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==2.3.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-reversion==3.0.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.0
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
Naked==0.1.31
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
pycparser==2.19
pycryptodome==3.7.0
PyJWT==1.6.4
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.6
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
shellescape==3.4.1
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24
xlrd==1.1.0
xlwt==1.3.0
